+++ src/DIC.php
@@ -30,23 +30,6 @@ use Typhoon\Type;
 final readonly class DIC
 {
     /**
-     * @template T
-     * @param callable(self): Ref<T> $app
-     * @return T
-     */
-    #[\Deprecated('Use DIC::init() instead', since: '0.3.3')]
-    public static function install(callable $app): mixed
-    {
-        /** @var T */
-        return Container::assemble(
-            static fn(Subscriber $subscriber, Tagger $tagger) => $app(new self(
-                subscriber: $subscriber,
-                tagger: $tagger,
-            )),
-        );
-    }
-
-    /**
      * Assembles the container and resolves the refs returned by $app.
      *
      * @param callable(self): mixed $app